Prevent cleaning released volumes more than once
There is a small time window between the deleter finding a successfully
completed cleanup job and the cache entry being removed in which
DeletePVs() may be called again. If this happens, then a second cleanup
job may be started before the cache entry is removed. This fix prevents
starting another cleanup job for a cache entry that has already been
cleaned successfully.