feat: support applying tolerations to cleanup jobs
At the moment, if the DaemonSet is deployed with tolerations, and `useJobForCleaning: true`, those tolerations do not get transferred to deployed Jobs. If, for example, the DaemonSet tolerates running on control plane nodes, the cleanup jobs it deploys targeting those nodes will never get scheduled due to lack of toleration. By adding a tolerations field to the config, the Helm chart can automatically populate it based on `.Values.tolerations`. Now all deployed jobs should inherit the same tolerations as the parent DaemonSet.