fix: Do not clean PV's with a deletion timestamp
After the volume has been successfully cleaned, the deleter sends a Delete
request to the API server. It should not continue attempting to clean the PV
at this point if the object still exists, as the PV may disappear at any time,
and we don't want cleanup to be in progress when that happens.